ecore-drm2: Add API function to retrieve output user data
With the rework of Ecore_Evas drm engine, we need to get the
ecore_evas itself in the pageflip callback so we have to reassign
output->user_data to the ecore_evas which we can then retrieve via
this function.

This is needed as the pageflip callback now passes us the output from
which the pageflip is complete.

@feature

Signed-off-by: Chris Michael <cp.michael@samsung.com>